+++ eventually-core/src/versioning.rs
@@ -4,7 +4,7 @@
 /// Data type that carries a version for Optimistic Concurrency Control.
 pub trait Versioned {
     /// Current version of the data.
-    fn version(&self) -> u32;
+    fn version(&self) -> i64;
 }
 
 impl<T> Versioned for Option<T>
@@ -12,7 +12,7 @@ where
     T: Versioned,
 {
     #[inline]
-    fn version(&self) -> u32 {
+    fn version(&self) -> i64 {
         self.as_ref().map(Versioned::version).unwrap_or_default()
     }
 }
